+++ gcpy/benchmark/modules/benchmark_utils.py
@@ -4,6 +4,7 @@ TODO: Migrate other benchmark-specific utilities from gcpy/benchmark.py to here.
 """
 import os
 import numpy as np
+import pandas as pd
 from gcpy import util
 from gcpy.constants import skip_these_vars
 
@@ -263,3 +264,52 @@ def print_benchmark_info(
         print(" - GCHP vs GCHP")
     if conf["gchp_vs_gcc_diff_of_diffs"]["run"]:
         print(" - GCHP vs GCC diff of diffs")
+
+
+def get_geoschem_level_metadata(
+        filename=None,
+        search_key=None,
+        verbose=False,
+):
+    """
+    Reads a comma-separated variable (.csv) file with GEOS-Chem vertical
+    level metadata and returns it in a pandas.DataFrame object.
+
+    Args:
+    -----
+    filename : str
+        Name of the comma-separated variable to read.
+        Default value: "__file__/GC_72_vertical_levels.csv"
+
+    Keyword Args:
+    -------------
+    search_key : str
+        If present, will return metadata that matches this value.
+        Default: None
+
+    verbose : bool
+        Toggles verbose printout on (True) or off (False).
+        Default value: True
+
+    Returns:
+    --------
+    metadata : pandas.DataFrame
+        Metadata for each of the GEOS-Chem vertical levels.
+    """
+    if filename is None:
+        filename = os.path.join(
+            os.path.dirname(__file__),
+            "GC_72_vertical_levels.csv"
+        )
+
+    try:
+        if verbose:
+            print(f"get_geoschem_level_metadata: Reading {filename}")
+        metadata = pd.read_csv(filename)
+    except (IOError, OSError, FileNotFoundError) as exc:
+        msg = f"Could not read GEOS-Chem level metadata in {filename}!"
+        raise exc(msg) from exc
+
+    if search_key is None:
+        return metadata
+    return metadata[search_key]